Updated handler factory

+++ backend/controllers/handlerFactory.js
@@ -5,7 +5,10 @@ exports.updateOne = (Model) =>
   catchAsync(async (req, res, next) => {
     let id;
     if (req.params.reviewId) id = req.params.reviewId;
-    else if (req.params.tourId) id = req.params.reviewId;
+    else if (req.params.tourId) id = req.params.tourId;
+    else if (req.user) id = req.user.id;
+
+    if (!id) return next(new AppError('Document ID is undefined', 500));
 
     const doc = await Model.findByIdAndUpdate(id, req.body, {
       new: true,
@@ -16,6 +19,7 @@ exports.updateOne = (Model) =>
 
     res.status(200).json({
       status: 'success',
+      data: doc,
     });
   });
 
@@ -23,7 +27,10 @@ exports.deleteOne = (Model) =>
   catchAsync(async (req, res, next) => {
     let id;
     if (req.params.reviewId) id = req.params.reviewId;
-    else if (req.params.tourId) id = req.params.reviewId;
+    else if (req.params.tourId) id = req.params.tourId;
+    else if (req.user) id = req.user.id;
+
+    if (!id) return next(new AppError('Document ID is undefined', 500));
 
     const doc = await Model.findByIdAndDelete(id);
     if (!doc) return next(new AppError('Cannot find document', 404));
